agentik-b2b-app/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.3.0
redis==5.0.1
celery==5.3.4
pydantic==2.5.1
pydantic-settings==2.2.1
python-dotenv==1.0.0
PyJWT==2.8.0
requests==2.31.0
aiofiles==23.2.1
python-multipart==0.0.6
email-validator==2.1.0
jinja2==3.1.2
aiosmtplib==3.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
aioredis==2.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
loguru==0.7.2
pytest==8.2.0
black==24.3.0
ruff==0.5.5
pyotp==2.9.0
